third_party/blink/web_tests/external/wpt/css/css-box/margin-trim/computed-margin-values/grid-inline-start-expected.txt

This is a testharness.js-based test.
[FAIL] grid > item 1
  assert_equals: \n<item data-expected-margin-left="0"></item>\nmargin-left expected "0" but got "30"
[FAIL] grid > item 3
  assert_equals: \n<item data-expected-margin-left="0"></item>\nmargin-left expected "0" but got "8"
[FAIL] grid > item 4
  assert_equals: \n<item class="locked-position" data-expected-margin-left="0"></item>\nmargin-left expected "0" but got "-30"
[FAIL] grid > item 5
  assert_equals: \n<item data-expected-margin-left="0"></item>\nmargin-left expected "0" but got "10"
Harness: the test ran to completion.
